backend/pyproject.toml
[project]
name = "cer-agent"
version = "0.1.0"
description = ""
authors = []
requires-python = "~=3.11"
readme = "README.md"
dependencies = [
    "django~=5.2",
    "django-cors-headers>=4.7.0,<5",
    "djangorestframework>=3.16.0,<4",
    "djangorestframework-simplejwt>=5.5.0,<6",
    "concurrent-log-handler>=0.9.25,<0.10",
    "python-dotenv>=1.1.0,<2",
    "psycopg-binary>=3.2.6,<4",
    "psycopg2>=2.9.10,<3",
    "gunicorn>=23.0.0,<24",
    "uvicorn>=0.34.0,<0.35",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "google-genai>=1.0.0,<2",
    "langgraph>=0.2.0,<0.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0,<26",
    "ruff>=0.11.4,<0.12",
]

[tool.black]
include = '\.pyi?$'
extend-exclude = '''
(
    \.git/
    | \.ruff_cache/
    | logs/
    | media/
    | static/
    | venv/
    | .venv/
)
'''
line-length = 100

[tool.ruff]
lint.select = [
    "Q",
    "I",
    "F401",
]
lint.exclude = [
    ".git",
    ".ruff_cache",
    "logs",
    "media",
    "static",
    "venv",
    ".venv",
]
line-length = 100

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"
docstring-quotes = "double"
